requirements.txt
Flask
google-generativeai
reportlab
gunicorn
python-dotenv

